[type] Migrate AccountStorage to account_storage, TableInfo to raw_table.
