+++ moveos/moveos-types/src/move_types.rs
@@ -1,8 +1,6 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::addresses::MOVEOS_STD_ADDRESS;
-use crate::moveos_std::object;
 use anyhow::{anyhow, bail, Result};
 use move_core_types::{
     account_address::AccountAddress, identifier::Identifier, language_storage::ModuleId,
@@ -206,12 +204,6 @@ pub fn type_tag_prop_strategy() -> impl Strategy<Value = TypeTag> {
     );
 
     type_tag_strategy
-}
-
-pub fn is_table(struct_tag: &StructTag) -> bool {
-    struct_tag.address == MOVEOS_STD_ADDRESS
-        && struct_tag.module.as_ident_str() == object::TABLE_INFO_MODULE_NAME
-        && struct_tag.name.as_ident_str() == object::TABLE_INFO_STRUCT_NAME
 }
 
 struct IdentifierSymbols;
