+++ moveos/moveos-types/src/moveos_std/mod.rs
@@ -1,6 +1,7 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod account_storage;
 pub mod any;
 /// This mod contains all the Rust to Move type mapping that are used in MoveosStd
 pub mod context;
@@ -10,6 +11,7 @@ pub mod module_upgrade_flag;
 pub mod move_module;
 pub mod object;
 pub mod object_ref;
+pub mod raw_table;
 pub mod simple_map;
 pub mod tx_context;
 pub mod tx_meta;
